Moved FrameworkHelper constructor tests to their own file.
